+++ src/preconditions/OwnerOnly.ts
@@ -1,6 +1,6 @@
 import { Precondition } from "@sapphire/framework";
 import type { CommandInteraction } from "discord.js";
-import { owner } from "../config/config.json";
+import { ownerId } from "../config/config.json";
 
 export class OwnerOnlyPrecondition extends Precondition {
   public override async chatInputRun(interaction: CommandInteraction) {
@@ -8,7 +8,7 @@ export class OwnerOnlyPrecondition extends Precondition {
   }
 
   private async checkOwner(userId: string) {
-    return owner === userId
+    return ownerId === userId
       ? this.ok()
       : this.error({ message: "Only the bot owner can use this command." });
   }
